Add a log handler
Currently only two log levels are supported:
 - INFO:  Normal messages
 - ERROR: Error messages

When the code is build without NDEBUG the function name and the line
number of the caller are printed. Also ERROR messages are echoed to
stderr. Before the log file handler can be used it must be initialzed
with initlog(). Initializing rotates the existing log files. When the
program is terminated the log handler must be closed with closelog().