Implement a simple command history
This is the approach used by plain bash, libedit, libreadline and so on.
The history can be search upwards and downwards, one step at a time. If
a user enters a command twice, it's saved twice.

A more sophisticated system with incremental search und duplicate remove
would be nice, but is much more complex. So maybe at a later time.